+++ inventory/src/main/java/com/netcracker/sharlan/bean/PaymentStatus.java
@@ -3,13 +3,13 @@ package com.netcracker.sharlan.bean;
 public enum PaymentStatus {
     NONE(0), BILLED(1), PAID(2), CANCELED(10);
 
-    private int id;
+    private long id;
 
-    PaymentStatus(int id){
+    PaymentStatus(long id){
         this.id = id;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
